refactor: split monolithic cni.go into feature-aligned files
Break cni.go into multiple files organized by feature for easier
reviews.  cni.go remains the package entry point.  Add package
documentation

- Extract config parsing and annotation key helpers into config.go
- Move resourceTracker and selective rollback into resource.go
- Extract BGP config, CRD lifecycle, and SRv6 setup into bgp.go
- Extract network namespace helpers into netns.go
- Extract CNI result builders into result.go
- Extract IPAM configure and deallocate operations into ipam_ops.go
- Extract host-device plugin delegation into host_device.go
- Extract cmdCheck and cmdStatus into ops_check.go
- Extract cmdAdd into ops_add.go
- Extract cmdDel into ops_del.go
- Add doc.go package godoc for the cni package
- Extend types.go with PluginConf, ipamResult, HostDevicePluginConf
