+++ internal/cni/netns.go
@@ -1,0 +1,117 @@
+// Copyright 2025 Datum Cloud, Inc.
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+package cni
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/containernetworking/plugins/pkg/ns"
+	"github.com/vishvananda/netlink"
+)
+
+// configureInterfaceInNetns applies an IP address and routes to the guest
+// interface inside the container network namespace.
+func configureInterfaceInNetns(netnsPath, ifName string, ipNet *net.IPNet, gateway net.IP) error {
+	containerNS, err := ns.GetNS(netnsPath)
+	if err != nil {
+		return fmt.Errorf("get container netns %q: %w", netnsPath, err)
+	}
+	defer containerNS.Close() //nolint:errcheck // netns close on teardown
+
+	return containerNS.Do(func(_ ns.NetNS) error {
+		handle, err := netlink.NewHandle()
+		if err != nil {
+			return fmt.Errorf("create netlink handle: %w", err)
+		}
+		defer handle.Close() //nolint:errcheck // netlink cleanup on teardown
+
+		link, err := handle.LinkByName(ifName)
+		if err != nil {
+			return fmt.Errorf("find guest interface %q: %w", ifName, err)
+		}
+
+		if err := handle.AddrAdd(link, &netlink.Addr{IPNet: ipNet}); err != nil {
+			return fmt.Errorf("add IP %s to %q: %w", ipNet, ifName, err)
+		}
+
+		if err := handle.LinkSetUp(link); err != nil {
+			return fmt.Errorf("set interface %q up: %w", ifName, err)
+		}
+
+		// Install default route via gateway.
+		if gateway != nil {
+			defaultRoute := &netlink.Route{
+				Dst:       nil, // default route
+				Gw:        gateway,
+				LinkIndex: link.Attrs().Index,
+			}
+			if err := handle.RouteAdd(defaultRoute); err != nil {
+				return fmt.Errorf("add default route via %s: %w", gateway, err)
+			}
+		}
+
+		return nil
+	})
+}
+
+// readGuestInterface reads the MAC and MTU of the guest veth endpoint
+// inside the container network namespace.
+func readGuestInterface(netnsPath, ifName string) (string, int, error) {
+	containerNS, err := ns.GetNS(netnsPath)
+	if err != nil {
+		return "", 0, fmt.Errorf("open container netns %s: %w", netnsPath, err)
+	}
+	defer containerNS.Close() //nolint:errcheck // netns close on teardown
+
+	var mac string
+	var mtu int
+	err = containerNS.Do(func(_ ns.NetNS) error {
+		handle, err := netlink.NewHandle()
+		if err != nil {
+			return fmt.Errorf("create netlink handle: %w", err)
+		}
+		defer handle.Close() //nolint:errcheck // netlink cleanup on teardown
+
+		link, err := handle.LinkByName(ifName)
+		if err != nil {
+			return fmt.Errorf("find interface %s: %w", ifName, err)
+		}
+		attrs := link.Attrs()
+		mac = attrs.HardwareAddr.String()
+		mtu = attrs.MTU
+		return nil
+	})
+	return mac, mtu, err
+}
+
+// cleanupContainerNetns removes any existing interface with the given name
+// from the container network namespace. This is needed to handle stale state
+// from previous CNI ADD runs that may have left interfaces behind.
+func cleanupContainerNetns(netnsPath, ifName string) error {
+	containerNS, err := ns.GetNS(netnsPath)
+	if err != nil {
+		return fmt.Errorf("get container netns %q: %w", netnsPath, err)
+	}
+	defer containerNS.Close() //nolint:errcheck // netns close on teardown
+
+	return containerNS.Do(func(_ ns.NetNS) error {
+		handle, err := netlink.NewHandle()
+		if err != nil {
+			return fmt.Errorf("create netlink handle: %w", err)
+		}
+		defer handle.Close() //nolint:errcheck // netlink cleanup on teardown
+
+		link, err := handle.LinkByName(ifName)
+		if err != nil {
+			// Interface does not exist in container netns — nothing to clean up.
+			return nil
+		}
+		if err := handle.LinkDel(link); err != nil {
+			return fmt.Errorf("delete stale interface %q in container netns: %w", ifName, err)
+		}
+		return nil
+	})
+}
