+++ internal/cni/types.go
@@ -4,11 +4,20 @@
 
 package cni
 
+import (
+	"net"
+
+	"github.com/containernetworking/cni/pkg/types"
+)
+
+// Termination represents a network termination point with a destination
+// CIDR and next-hop gateway address.
 type Termination struct {
 	Network string `json:"network"`
 	Via     string `json:"via,omitempty"`
 }
 
+// IPAM holds IP address management configuration passed in the CNI config.
 type IPAM struct {
 	Type      string    `json:"type"`                 // "pool" (default) or "static"
 	Pool      string    `json:"pool,omitempty"`       // IPv6 CIDR pool, e.g. "fd00:10:ff01::/48"
@@ -19,11 +28,40 @@ type IPAM struct {
 	Addresses []Address `json:"addresses,omitempty"`
 }
 
+// Route describes a static route to install.
 type Route struct {
 	Dst string `json:"dst"`
 	GW  string `json:"gw,omitempty"`
 }
 
+// Address describes a static IP address assignment.
 type Address struct {
 	Address string `json:"address"`
+}
+
+// PluginConf is the CNI plugin configuration passed via stdin on each invocation.
+type PluginConf struct {
+	types.PluginConf
+	VPC           string        `json:"vpc"`
+	VPCAttachment string        `json:"vpcattachment"`
+	MTU           int           `json:"mtu,omitempty"`
+	InterfaceType string        `json:"interface_type,omitempty"` // interfaceTypeVeth or interfaceTypeTap
+	Terminations  []Termination `json:"terminations,omitempty"`
+	IPAM          IPAM          `json:"ipam"`
+	SRv6Locator   string        `json:"srv6_locator"`
+	Namespace     string        `json:"namespace,omitempty"`
+}
+
+// ipamResult holds the IPAM allocation details for building the CNI result.
+type ipamResult struct {
+	subnet  *net.IPNet
+	gateway net.IP
+	routes  []*net.IPNet
+}
+
+// HostDevicePluginConf is the configuration for the host-device CNI plugin
+// delegation used to move the guest veth endpoint into the container netns.
+type HostDevicePluginConf struct {
+	types.PluginConf
+	Device string `json:"device"`
 }
